% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{attributes<-.functions}
\alias{attributes<-.functions}
\title{Prevent attribute renaming in class functions}
\usage{
attributes.functions(object) <- value
}
\description{
Prevent attribute renaming in class functions
}
