% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{attributes<-.themetadata}
\alias{attributes<-.themetadata}
\title{Prevent attribute renaming in class themetadata}
\usage{
attributes.themetadata(object) <- value
}
\description{
Prevent attribute renaming in class themetadata
}
