% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_topics.R
\name{attributes<-.topics}
\alias{attributes<-.topics}
\title{Prevent attribute renaming in class topics}
\usage{
attributes.topics(object) <- value
}
\description{
Prevent attribute renaming in class topics
}
