% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_for_splines}
\alias{check_for_splines}
\title{Checks whether a continuous covariate involves a basis function}
\usage{
check_for_splines(formula, metadata)
}
\description{
Checks whether a continuous covariate involves a basis function
}
