% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn.R
\name{cnn}
\alias{cnn}
\title{Normalize an OTU table by 16S rRNA copy number}
\usage{
cnn(otu_table, rows_are_taxa, copy_numbers, drop = TRUE)
}
\arguments{
\item{otu_table}{Phyloseq object or OTU table as a data frame or matrix with
OTU IDs as row or column names.}

\item{rows_are_taxa}{TRUE/FALSE whether OTUs are rows and samples are columns
or vice versa.}

\item{copy_numbers}{A 2-column matrix or data frame of copy numbers where
column 1 contains the OTU IDs and column 2 the copy numbers.}

\item{drop}{Logical whether to drop empty samples or OTUs after normalization.
Defaults to TRUE.}
}
\value{
A normalized, rounded (to nearest integer) OTU table.
}
\description{
This implements copy number normalization using the PICRUSt 16S GreenGreenes
13.5 copy number count table (default) or a user provided set of copy
numbers.
}
