% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_modelframe}
\alias{create_modelframe}
\title{Create modelframe that expands catagoricals into dummy variables}
\usage{
create_modelframe(formula, refs, metadata)
}
\description{
Create modelframe that expands catagoricals into dummy variables
}
