% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_multiclasses_table}
\alias{create_multiclasses_table}
\title{Create a lookup table that identifies splines and multiclass factors}
\usage{
create_multiclasses_table(modelframe, modelframe_full, splines = NULL)
}
\description{
Create a lookup table that identifies splines and multiclass factors
}
