% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ref.R
\name{download_ref}
\alias{download_ref}
\title{Download PICRUSt reference tables}
\usage{
download_ref(reference, destination, verbose = FALSE)
}
\arguments{
\item{reference}{A string: ko, cog, or all.}

\item{destination}{Location of the folder to save the reference files.}

\item{verbose}{Logical to print status of download.}
}
\description{
A function to download the KO and COG 13.5 GreenGenes reference tables
for PICRUSt prediction.
}
