% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{est}
\alias{est}
\title{Estimate topic or function effects}
\usage{
est(object, ...)
}
\arguments{
\item{object}{A topics or functions object generated by \code{\link{find_topics}} or \code{\link{topics.predict}}, respectively.}

\item{...}{Additional arguments for methods.}
}
\description{
Estimate topic or function effects
}
\examples{
topics <- find_topics(K=30,otu_table=otu_table,rows_are_taxa=FALSE,formula=group + time,metadata=metadata,refs=c('A'))
topic_effects <- est(topics,metadata=metadata,formula=~group + time,refs=c('A'),moderator='group')

download_ref('~/references')
functions <- predict(topics,reference_path='~/references/ko_13_5_precalculated.tab.gz',scalar=100)
function_effects <- est(functions,level=2,method='HMC')
}
