% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_topic_effects.R
\name{est.topics_backend}
\alias{est.topics_backend}
\title{Backend to extract effects for \code{\link{estimate_topic_effects}}.}
\usage{
\method{est}{topics_backend}(estimated_effects, theta, nsims = 100,
  ui_level = 0.8, npoints = 100, verbose = FALSE)
}
\description{
Backend to extract effects for \code{\link{estimate_topic_effects}}.
}
\keyword{internal}
