% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_function_effects.R
\name{estimate_function_effects}
\alias{estimate_function_effects}
\title{Estimate the predicted function-topic effects}
\usage{
estimate_function_effects(functions, topics_subset, level = 2,
  method = "ML", ...)
}
\arguments{
\item{functions}{List containing the output of
\code{\link{predict_functions}}.}

\item{topics_subset}{Vector of intergers corresponding to a subset of topics
to be evaluated. Recommended to be < 25.}

\item{level}{(optional) The gene category level to evalulate. Defaults to 2.}

\item{method}{(optional) String indicating either ML or HMC. Defaults to ML.}

\item{verbose}{(optional) Logical to print progress. Defaults to TRUE.}
}
\value{
A glmerMod object (ML) or a list containing rstan output
  (HMC).
}
\description{
Given within topic functional predictions, estimate the effects at a given
gene function category level. The effects correspond to a topic-gene category
interaction term after accounting for topic and gene category effects. The
model can be fit via either maximum likelihood or Hamiltonian MC -- see
details.
}
