% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_topic_effects.R
\name{estimate_topic_effects}
\alias{estimate_topic_effects}
\title{Estimate topic effects for a given covariate}
\usage{
estimate_topic_effects(covariate, topics, metadata, nsims = 100,
  ui_level = 0.8)
}
\arguments{
\item{covariate}{String corresponding to the name of the covariate of interest found in metadata.}

\item{topics}{STM object containing the topic information}

\item{metadata}{Dataframe or matrix of metadata. If the STM was fit with covariate information, this must point to the same metadata used in the STM.}

\item{nsims}{(optional) Number of simulations to perform for inferring covariate effects.}

\item{ui_level}{With of uncertainty interval to report effects. Defaults to .95.}
}
\value{
A list containing

\item{est}{The estimate of the effect}
\item{rank}{The rank of the topics based on the effect estimate (increasing)}
\item{sig}{A vector of topic indexes for topics whose uncertainty intervals did not enclose 0.}
}
\description{
Given a covariate of interest, measure its relationship with the samples over topics distribution estimated by the STM.
}
