% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_effects.R
\name{extract}
\alias{extract}
\alias{extract.effects}
\title{Extract summary statistics}
\usage{
extract(object, ...)

\method{extract}{effects}(object, verbose = FALSE)
}
\value{
A list containing
\describe{
\item{summary}{Rstan summary of parameters from model.}
\item{flagged}{Vector of parameter names with Rhat > 1.1.}
}
}
\description{
Extract summary statistics
}
\section{Methods (by class)}{
\itemize{
\item \code{effects}: Extract summary statistics from HMC effects object

Extracts the summary information in a form condusive with vis methods,
specifically in cases when return_summary was set to FALSE.
}}

