% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_spline_info}
\alias{extract_spline_info}
\title{Extract basis function design matrix and basis type}
\usage{
extract_spline_info(formula, metadata, remove_only = FALSE)
}
\description{
Extract basis function design matrix and basis type
}
