% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_effects.R
\name{extract_stan_summary}
\alias{extract_stan_summary}
\title{Backend to extract summary from rstan object}
\usage{
extract_stan_summary(fit, stan_dat, summary_pars)
}
\description{
Backend to extract summary from rstan object
}
\keyword{internal}
