% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_topics.R
\name{find_topics}
\alias{find_topics}
\title{Fit a topic model on an OTU table}
\usage{
find_topics(K, otu_table, rows_are_taxa, formula, metadata, refs,
  control = list(), ...)
}
\arguments{
\item{K}{A positive integer indicating the number of topics to be estimated.}

\item{otu_table}{Phyloseq object or OTU table as a data frame or matrix with
OTU IDs as row or column names.}

\item{rows_are_taxa}{Logical whether OTUs are rows and samples are columns
or vice versa.}

\item{formula}{(optional) Formula object with no response variable or a
matrix containing topic prevalence covariates. Must provide metadata.}

\item{metadata}{Matrix or dataframe containing the metadata that corresponds
with the formula.}

\item{control}{(optional) List of additional parameters control portions of
the optimization. See details.}

\item{sigma_prior}{(optional) Scalar between 0 and 1. This sets the strength
of regularization towards a diagonalized covariance matrix. Setting the
value above 0 can be useful if topics are becoming too highly correlated.}

\item{model}{(optional) Prefit STM model object to restart an existing model.}

\item{iters}{(optional) Maximum number of EM iterations.}

\item{tol}{(optional) Convergence tolerance.}

\item{batches}{(optional) Number of groups for memoized inference.}

\item{seed}{(optional) Seed for the random number generator to reproduce
previous results.}

\item{verbose}{(optional) Logical indicating whether information should be
printed.}

\item{verbose_n}{(optional) Integer determining the intervals at which labels
are printed.}
}
\value{
A list containing

\item{fit}{The topic model fit}
\item{docs}{The documents}
\item{vocab}{The vocabulary}
}
\description{
Given an OTU table, this function converts the OTU counts across samples into
a document format and then fits a structural topic model by wrapping the stm
command from the STM package, see \code{\link{stm}}.
}
