% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ml_model.R
\name{fit_ml_model}
\alias{fit_ml_model}
\title{Estimate function effects via ML}
\usage{
fit_ml_model(gene_table, iters = 1000, return_fit = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{gene_table}{A gene_table formatted via \code{\link{format_gene_table}}.}

\item{iters}{Number of iterations for HMC. Defaults to 1000.}

\item{return_fit}{(optional) Logical to return the rstan data and fit.}

\item{verbose}{(optional) Logical to print progress.}

\item{inits}{List of values for parameter initialization. If omitted, values
are generated via \code{\link{lme4::glmer.nb}}}

\item{chains}{(optional) Number of chains for HMC. Defaults to 1.}
}
\value{
A list containing

\item{summary}{Rstan output that includes the coefficient mean estimates, uncertainty
intervals, standard deviations, effective sample size, Rhat statistics}
\item{fit}{Rstan fit (if return_fit=TRUE)}
\item{data}{Rstan input data (if return_fit=TRUE)}
\item{flagged}{Parameter estimates with Rhat statistics > 1.1}
\item{inits}{If parameter estimates are flagged, a list of inits to initialize Rstan
with more iterations.}
}
\description{
Given within topic functional predictions, estimate the effects at a given
gene function category level via ML The effects correspond to a topic-gene
category interaction term after accounting for topic and gene category
effects.
}
