% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_gene_metadata}
\alias{format_gene_metadata}
\title{functional metadata strings to lists of lists}
\usage{
format_gene_metadata(picrust_out)
}
\description{
functional metadata strings to lists of lists
}
