% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_gene_table}
\alias{format_gene_table}
\title{format functional output to gene table for inference}
\usage{
format_gene_table(functions, level, pw_targets, keep)
}
\description{
format functional output to gene table for inference
}
