% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_functions.R
\name{names<-.functions}
\alias{names<-.functions}
\title{Prevent object renaming in class functions}
\usage{
\method{names}{functions}(object) <- value
}
\description{
Prevent object renaming in class functions
}
