% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{names<-.themetadata}
\alias{names<-.themetadata}
\title{Prevent object renaming in class themetadata}
\usage{
\method{names}{themetadata}(object) <- value
}
\description{
Prevent object renaming in class themetadata
}
