% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{picrust}
\alias{picrust}
\title{Predict functional potential}
\usage{
picrust(file_path, otu_id_targets)
}
\arguments{
\item{file_path}{Path to the precalculated table}

\item{otu_id_targets}{Character vector of OTU IDs to predict}
}
\value{
A list containing the the gene IDs, the OTU IDs that
were present in the precalculated table and hence predicted,
the PICRUSt gene metadata, the predicted gene table, and the
PICRUSt metadata that contains confidence scores, etc.
}
\description{
This predicts the functional potential of a topic metagenome.
Raw OTU tables may be used as well. By default, the function
uses PICRUSt's precalculated GreenGreens table. If the user
chooses to provide his or her own, then the table must be
formatted accordingly.
}
