% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{picrust_otu}
\alias{picrust_otu}
\title{Predict OTU functional potential}
\usage{
picrust_otu(file_path, otu_id_targets)
}
\arguments{
\item{file_path}{Path to the precalculated table}

\item{otu_id_targets}{Character vector of OTU IDs to predict}
}
\value{
A list containing the the gene IDs, the OTU IDs that
were present in the precalculated table and hence predicted,
the PICRUSt gene metadata, the predicted gene table, and the
PICRUSt metadata that contains confidence scores, etc.
}
\description{
This predicts the functional potential of a metagenome given
a precalculated mapping table which maps the functional
content for a given OTU to content across functional genes.
}
