% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_functions.R
\name{predict_functions}
\alias{predict_functions}
\title{Predict topic functional content}
\usage{
predict_functions(fit, reference_path, scaler = 1000)
}
\arguments{
\item{fit}{STM object}

\item{scaler}{Value for scaling the topics over OTUs distrubution to
predicted counts. Defaults at 1000.}

\item{reference}{Location of the precalculated mapping table. Can be .gz.}
}
\value{
A list containing a matrix of gene counts across topics, a list
of associated functional metadata, and matrix of PICRUSt gene metadata.
}
\description{
Given an STM object, this function predicts the functional content using
a precalculated mapping table which maps the taxonomic abundance for a
given OTU to functional abundance content across a set of functional genes.
}
