% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for pipeline}
\usage{
prepare_data(otu_table, rows_are_taxa, taxa, metadata, formula, drop = TRUE)
}
\arguments{
\item{otu_table}{OTU table.}

\item{rows_are_taxa}{Logical flag.}

\item{taxa}{Taxa table.}

\item{metadata}{Metadata.}

\item{formula}{Formula for covariates.}

\item{drop}{Drop empty.}
}
\description{
TBD
}
