% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{prev_rename}
\alias{prev_rename}
\alias{names<-.topics}
\alias{attributes<-.topics}
\alias{names<-.themetadata}
\alias{attributes<-.themetadata}
\alias{names<-.functions}
\alias{attributes<-.functions}
\alias{names<-.effects}
\alias{attributes<-.effects}
\title{Prevent x renaming}
\usage{
\method{names}{topics}(x) <- value

attributes.topics(x) <- value

\method{names}{themetadata}(x) <- value

attributes.themetadata(x) <- value

\method{names}{functions}(x) <- value

attributes.functions(x) <- value

\method{names}{effects}(x) <- value

attributes.effects(x) <- value
}
\arguments{
\item{x}{x}

\item{value}{value}
}
\description{
Prevent x renaming
}
