% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.effects}
\alias{print.effects}
\title{Print summary for effects class resulting from topics}
\usage{
\method{print}{effects}(x, ...)
}
\description{
Print summary for effects class resulting from topics
}
