% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.functions}
\alias{print.functions}
\title{Print summary for effects class resulting from functions}
\usage{
\method{print}{functions}(functions_object, ...)
}
\description{
Print summary for effects class resulting from functions
}
