% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{print_effects_functions}
\alias{print_effects_functions}
\title{Print summary for effects class resulting from functions}
\usage{
print_effects_functions(effects_object)
}
\description{
Print summary for effects class resulting from functions
}
\keyword{internal}
