% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{print_effects_topics}
\alias{print_effects_topics}
\title{Print summary for effects class resulting from topics}
\usage{
print_effects_topics(effects_object)
}
\description{
Print summary for effects class resulting from topics
}
\keyword{internal}
