% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{rename_taxa_to_other}
\alias{rename_taxa_to_other}
\title{rename taxa below rank to other}
\usage{
rename_taxa_to_other(x, taxa, top_n = 7, group = c("Phylum", "Class",
  "Order", "Family", "Genus"), type = c("otu_table", "docs"),
  as_factor = FALSE)
}
\description{
rename taxa below rank to other
}
