% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_topics.R
\name{stm_wrapper}
\alias{stm_wrapper}
\title{Wrapper for \code{\link{stm}}}
\usage{
stm_wrapper(K, docs, vocab, formula = NULL, metadata = NULL,
  sigma_prior = 0, model = NULL, iters = 500, tol = 1e-05,
  batches = 1, seed = NULL, verbose = FALSE, verbose_n = 5,
  control = control)
}
\description{
Wrapper for \code{\link{stm}}
}
\keyword{internal}
