% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{sum_taxa_by_group}
\alias{sum_taxa_by_group}
\title{group_by summarize function for taxa without dplyr}
\usage{
sum_taxa_by_group(otu_ids, taxa, otu_table, metadata, cov_list,
  group = c("Phylum", "Class", "Order", "Family", "Genus"),
  sample_norm = FALSE)
}
\description{
group_by summarize function for taxa without dplyr
}
