% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_covariate_effect_continuous.R
\name{vis_covariate_effects_continuous}
\alias{vis_covariate_effects_continuous}
\title{Generate interactive continuous covariate effect figure}
\usage{
vis_covariate_effects_continuous(topics, topic_effects, otu_table, taxa,
  covariate, metadata, sample_norm = FALSE, taxa_n = 10)
}
\arguments{
\item{topics}{Output of \code{\link{find_topics}} that contains the STM object.}

\item{topic_effects}{Output of \code{\link{estimate_topic_effects}} that contains regression weights.}

\item{taxa}{Dataframe or matrix containing the taxonomy information.}

\item{function_effects}{Output of \code{\link{estimate_function_effects}} that contains the results from either HMC or ML.}

\item{beta_min}{(optional) Minimum probability in topics over taxa distribution to set to 0.}

\item{gene_min}{(optional) Mininum count for gene set table.}
}
\description{
TBD
}
