% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_function_effects.R
\name{vis_function_effects}
\alias{vis_function_effects}
\title{Generate interactive functional effect heatmap}
\usage{
vis_function_effects(topics, topic_effects, function_effects, taxa,
  beta_min = 1e-05, gene_min = 10)
}
\arguments{
\item{topics}{Output of \code{\link{find_topics}} that contains the STM object.}

\item{topic_effects}{Output of \code{\link{estimate_topic_effects}} that contains regression weights.}

\item{function_effects}{Output of \code{\link{estimate_function_effects}} that contains the results from either HMC or ML.}

\item{taxa}{Dataframe or matrix containing the taxonomy information.}

\item{beta_min}{(optional) Minimum probability in topics over taxa distribution to set to 0.}

\item{gene_min}{(optional) Mininum count for gene set table.}
}
\description{
TBD
}
