% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_topic_corr.R
\name{vis_topic_corr}
\alias{vis_topic_corr}
\title{Visualize topic correlations.}
\usage{
vis_topic_corr(topics, topics_effects, method = "huge")
}
\arguments{
\item{topics}{Output of \code{\link{find_topics}} that contains the STM object.}

\item{topics_effects}{Output of \code{\link{estimate_topic_effects}}}

\item{method}{Correlation graph method. Defaults to huge.}
}
\description{
TBD
}
