% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_topics.R
\name{vis_topics}
\alias{vis_topics}
\title{Visualize principal components of topics over samples distribution.}
\usage{
vis_topics(topics, otu_table, taxa, taxa_n = 25)
}
\arguments{
\item{topics}{Output of \code{\link{find_topics}} that contains the STM object.}

\item{otu_table}{Taxonomy table used in STM.}

\item{taxa}{Dataframe or matrix containing the taxonomy information.}

\item{taxa_n}{(optional) Number of taxa to show in table. Defaults to 25.}
}
\description{
TBD
}
