% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ref.R
\name{download_ref}
\alias{download_ref}
\title{Download PICRUSt reference tables}
\usage{
download_ref(destination, reference = "all", overwrite = FALSE,
  verbose = FALSE)
}
\arguments{
\item{destination}{Location of the folder to save the reference files.}

\item{reference}{(optional) A string for either gg_ko, gg_cog, silva_ko, or
all. Defaults to all.}

\item{verbose}{(optional) Logical to print status of download. Defaults to
TRUE.}
}
\description{
A function to download the KO and COG 13.5 GreenGenes reference tables for
PICRUSt prediction.
}
