% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{est}
\alias{est}
\title{Estimate topic or function effects}
\usage{
est(object, ...)
}
\arguments{
\item{object}{A topics or functions object generated by \code{\link{find_topics}} or \code{\link{predict}}, respectively.}

\item{...}{Additional arguments for methods.}
}
\description{
Estimate topic or function effects
}
\seealso{
\code{\link{est.topics}} \code{\link{est.functions}} \code{\link{est.hmc}} \code{\link{est.ml}}
}
