% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_topics.R
\name{find_topics}
\alias{find_topics}
\title{Fit a topic model on an OTU table}
\usage{
find_topics(K, otu_table, rows_are_taxa, control = list(), ...)
}
\arguments{
\item{K}{A positive integer indicating the number of topics to
be estimated.}

\item{otu_table}{Phyloseq object or OTU table as a data frame
or matrix with OTU IDs as row or column names.}

\item{rows_are_taxa}{TRUE/FALSE whether OTUs are rows and
samples are columns or vice versa.}

\item{control}{List of additional parameters control portions of
the optimization. See details.}

\item{formula}{Formula object with no response variable or a
matrix containing topic prevalence covariates.}

\item{sigma_prior}{scalar between 0 and 1. This sets the strength
of regularization towards a diagonalized covariance matrix.
Setting the value above 0 can be useful if topics are becoming
too highly correlated.}

\item{model}{Prefit STM model object to restart an existing model.}

\item{iters}{Maximum number of EM iterations.}

\item{tol}{Convergence tolerance.}

\item{batches}{Number of groups for memoized inference.}

\item{seed}{Seed for the random number generator to reproduce
previous results.}

\item{verbose}{Logical indicating whether information should be
printed.}

\item{verbose_n}{Integer determining the intervals at which labels
are printed.}
}
\value{
A list containing the topic model fit, the documents, and
the vocabulary.
}
\description{
Given an OTU table, this function converts the OTU counts
across samples into a document format and then fits a
structural topic model by wrapping the stm command from the
STM package, see \code{\link{stm::stm}}.
}
