% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picrust.R
\name{picrust}
\alias{picrust}
\title{Predict topic functional content via PICRUSt}
\usage{
picrust(otu_table, rows_are_taxa, reference_path, drop = TRUE)
}
\arguments{
\item{rows_are_taxa}{TRUE/FALSE whether OTUs are rows and samples are columns
or vice versa.}

\item{reference_path}{Location of the precalculated mapping file. See
\code{link{download_ref}}}

\item{drop}{(optional) Logical whether to drop empty samples or OTUs after
normalization. Defaults to TRUE.}

\item{x}{An abundance table of samples verses OTUs}
}
\value{
A list containing

\item{fxn_table}{A matrix of gene counts across topics}
\item{fxn_meta}{A list of associated functional metadata}
\item{pi_meta}{matrix of PICRUSt metadata (e.g., NSTI)}
}
\description{
Given an OTU abundance table prepared with the GreenGenes reference database,
this function predicts the functional content using either COG or KO
precalculated mapping tables that map the taxonomic abundance for a given OTU
to functional abundance content across a set of functional genes.
}
