% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis}
\alias{vis}
\alias{vis.topics}
\title{Visualize effects object with Shiny}
\usage{
vis(effects_object, tax_table, type = c("topic", "binary", "continuous",
  "function"), ...)
}
\arguments{
\item{effects_object}{An object of class effects, generated by \code{\link{est}}.}

\item{tax_table}{Taxonomic table associated with otu_table used in \code{\link{find_topics}}.}

\item{type}{Type of visualization to perform.}

\item{...}{Additional arguments for methods.}
}
\description{
Visualize effects object with Shiny
}
\examples{
topics <- find_topics(K=30,otu_table=otu_table,rows_are_taxa=FALSE,formula=group + time,metadata=metadata,refs=c('A'))

topic_effects <- est(topics,metadata=metadata,formula=~group + time,refs=c('A'),moderator='group')

vis(topic_effects,taxa=taxa_table,type='topic')
vis(topic_effects,taxa=taxa_table,type='binary')
vis(topic_effects,taxa=taxa_table,type='continuous',moderator='group')
}
