% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_topic_features.R
\name{vis_topic_features}
\alias{vis_topic_features}
\title{Visualize principal components of topics over samples distribution.}
\usage{
vis_topic_features(topics, otu_table, topic_effects, taxa, lambda_step = 0.01,
  taxa_n = 30, top_n = 7, method = "huge")
}
\arguments{
\item{topics}{Output of \code{\link{find_topics}} that contains the STM object.}

\item{otu_table}{Taxonomy table used in STM.}

\item{taxa}{Dataframe or matrix containing the taxonomy information.}

\item{taxa_n}{(optional) Number of taxa to show in table. Defaults to 25.}
}
\description{
TBD
}
